requirements.txt
alembic==0.8.6
appnope==0.1.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
boto3==1.3.1
botocore==1.4.28
certifi==2016.2.28
click==6.6
decorator==4.0.10
docutils==0.12
Flask==0.11.1
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
futures==3.0.5
gnureadline==6.3.3
ipython==4.2.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
Mako==1.0.4
MarkupSafe==0.23
MySQL-python==1.2.5
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
PyMySQL==0.7.4
python-dateutil==2.5.3
python-editor==1.0.1
RPi.GPIO==0.6.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.0.13
tornado==4.3
traitlets==4.2.1
Werkzeug==0.11.10
